Replace psycopg2-binary with pg8000 and black with ruff (#25)
* Replace psycopg2-binary with pg8000

psycopg2 is licensed under LGPL that's not compatible with ASF -
https://www.apache.org/legal/resolved.html#category-x

* Replace black with ruff

black depends on pathspec that is licensed under MPL 2.0. Even though
MPL 2.0 can _maybe_ be included in an ASF project, I'd
prefer to simplify the discussion and just replace black with ruff that
is licensed as Apache 2.0, does not have any dependencies, and is
already used by other Apache projects.